LWSHADOOP-748: Look in HDFS Index Dir when deleting write.lock files
Ambari provides an option for users to have their write.lock files deleted
on startup, if they'd like.

Prior to this commit, this cleanup was hardcoded to look in the default
HDFS index dir: `/solr`.  This leaves lock files behind, and causes
problems for the Solr processes during or after startup.

This commit switches the cleanup over to using the user-specified
configuration option, instead of the hardcoded path.