LWSHADOOP-748: Only remove lock files associated with Solr node
Prior to this commit, the "Delete write.lock files on startup" option
would delete _all_ write.lock files it could find.  This behavior is
appropriate when the cluster is coming online for the first time, or
after a maintenance period.  However it's too aggressive for when a
single node in a larger cluster comes back online.

This commit makes our lock-file-deletion more granular, only deleting
lock files that are "owned" by the Solr node being started (according
to the collection state.json files in ZooKeeper).

Based on PR opened by vzak-profimedia here:
https://github.com/lucidworks/solr-stack/pull/28