Remove reference to deleted Solr configset
Prior versions of Solr included a "data_driven" example configset,
however this configset was removed during the 7.x release line and
is no longer present in 7.3.1.

The example-collection configuration referenced this configset, causing
the post-startup Solr Check to fail.  This commit changes the test
collection creation to use the "_default" configset instead.