Modify HDP stack requirements for 6.6.2
With the introduction of SOLR-7.3.1, we now have two versions of the
SOLR service that (according to their metainfo.xml and mpack.json
specs) are meant for HDP-2.6.  This commit modifies the service
metadata so that SOLR 5.5.5 is used on HDP 2.4, SOLR 6.6.2 is used on
HDP 2.5, and SOLR 7.3.1 is used on HDP 2.6.