LWSHADOOP-721: Use ticketCache for Solr->ZK Conn
Prior to this commit, when Solr was connecting to ZK with Kerberos
enabled, it would use a specified keytab only. It would not attempt
to connect via a ticket cache.

With this commit, Solr will attempt to use the ticket cache first,
before falliong back to explicitly using the keytab. For more
information, see:
https://docs.oracle.com/javase/8/docs/jre/api/security/jaas/spec/com/sun/security/auth/module/Krb5LoginModule.html