LWSHADOOP-773: Change JAAS profile name to default
Prior to this commit, the JAAS file for solr-metrics auth used a profile
called "SolrJClient".  But this non-default name was never used by
anything, which often caused ticket renewal and other Kerberos problems
to crash metrics.

This commit changes the JAAS file to use the default profile name
instead, so that SolrJ finds the right profile automatically.